uint32 internally

+++ examples/red_square_exit_on_esc_key.nim
@@ -5,4 +5,4 @@ var app = init(Fenster, "red square. exit on esc key", 800, 600)
 while app.loop() and app.keys[27] == 0:
   for x in 350 .. 450:
     for y in 250 .. 350:
-      app[x, y] = uint32(0xFF0000)
+      app[x, y] = 0xFF0000
